backend/requirements.txt
torch
torchvision
opencv-python-headless
numpy
gradio
pillow
